+++ lib/main.dart
@@ -2,9 +2,9 @@
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'views/HomeView.dart';
+import 'views/home_view.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:projectloner/auth/CheckLogin.dart';
+import 'package:projectloner/auth/check_login.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:projectloner/blocs/swipe/swipe_bloc.dart';
 import 'package:projectloner/matching/matching_screen.dart';
@@ -28,7 +28,7 @@ class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {  
+  Widget build(BuildContext context) {
     return MultiBlocProvider(
       providers: [
         BlocProvider(
