+++ lib/views/home_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:projectloner/views/matchViews/CoachView.dart';
-import 'package:projectloner/views/matchViews/DuoView.dart';
-import 'package:projectloner/views/matchViews/TeamView.dart';
+import 'package:projectloner/views/matchViews/coach_view.dart';
+import 'package:projectloner/views/matchViews/duo_view.dart';
+import 'package:projectloner/views/matchViews/team_view.dart';
 import '../matching/matching_screen.dart';
-import 'NavBar.dart';
+import 'nav_bar.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
@@ -13,15 +13,11 @@ class HomePage extends StatefulWidget {
 }
 
 class HomePageState extends State {
-
   String preferenceText = "Currently looking for duo";
   String matchButtonText = "Start";
   final List<bool> _matchPreferences = <bool>[true, false, false];
 
-
-  void updatePage() {
-    
-  }
+  void updatePage() {}
 
   void startMatching(List<bool> matchPreferences) {
     if (matchPreferences[0] == true) {
@@ -31,14 +27,14 @@ class HomePageState extends State {
           builder: (context) => const MatchMaking(),
         ),
       );
-    }
-    else if (matchPreferences[1] == true) {
+    } else if (matchPreferences[1] == true) {
       print("Team");
-      Navigator.push(context, MaterialPageRoute(builder: (context) => TeamPage()));
-    }
-    else if (matchPreferences[2] == true) {
+      Navigator.push(
+          context, MaterialPageRoute(builder: (context) => TeamPage()));
+    } else if (matchPreferences[2] == true) {
       print("Coach");
-      Navigator.push(context, MaterialPageRoute(builder: (context) => CoachPage()));
+      Navigator.push(
+          context, MaterialPageRoute(builder: (context) => CoachPage()));
     }
   }
 
@@ -47,15 +43,14 @@ class HomePageState extends State {
     return Scaffold(
       appBar: AppBar(title: const Text("Home Page")),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            const Text( 
+        child:
+            Column(mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: [
+          const Text(
             'Home Page',
             style: TextStyle(fontSize: 30.0, fontWeight: FontWeight.bold),
-            ),
-            // I am stoopid
-            /*
+          ),
+          // I am stoopid
+          /*
             ButtonBar(
               mainAxisSize: MainAxisSize.min,
               children: [
@@ -79,7 +74,7 @@ class HomePageState extends State {
                 ),
               ],
             ),*/
-            ToggleButtons(
+          ToggleButtons(
               onPressed: (int index) {
                 setState(() {
                   for (int i = 0; i < _matchPreferences.length; i++) {
@@ -93,22 +88,21 @@ class HomePageState extends State {
               selectedColor: Colors.white,
               fillColor: Colors.deepPurple,
               constraints: const BoxConstraints(
-                  minHeight: 40.0,
-                  minWidth: 80.0,
+                minHeight: 40.0,
+                minWidth: 80.0,
               ),
               children: const [
                 Text("Duo"),
                 Text("Team"),
                 Text("Coach"),
               ]),
-            //Text(preferenceText),
-            ElevatedButton(
+          //Text(preferenceText),
+          ElevatedButton(
               onPressed: () {
                 startMatching(_matchPreferences);
-              }, 
+              },
               child: Text(matchButtonText))
-          ]
-        ),
+        ]),
       ),
       drawer: const NavBar(),
     );
