+++ lib/views/matchViews/team_view.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:projectloner/views/NavBar.dart';
+import 'package:projectloner/views/nav_bar.dart';
 
-import '../HomeView.dart';
+import '../home_view.dart';
 
 class TeamPage extends StatefulWidget {
-  
   const TeamPage({Key? key}) : super(key: key);
 
   @override
@@ -12,27 +11,23 @@ class TeamPage extends StatefulWidget {
 }
 
 class TeamPageState extends State {
-  
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(title: Text("Team Matching")),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Text( 
+        child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+          const Text(
             'Team matchmaking here',
             style: TextStyle(fontSize: 30.0, fontWeight: FontWeight.bold),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pop(context);
-              }, 
-              child: const Text("Go Back"),
-            ),
-          ]
-        ),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.pop(context);
+            },
+            child: const Text("Go Back"),
+          ),
+        ]),
       ),
       drawer: const NavBar(),
     );
