+++ lib/views/nav_bar.dart
@@ -3,8 +3,8 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import "package:flutter/material.dart";
 import 'package:instabug_flutter/instabug_flutter.dart';
-import 'HomeView.dart';
-import 'package:projectloner/auth/AuthPage.dart';
+import 'home_view.dart';
+import 'package:projectloner/auth/auth_page.dart';
 import 'package:projectloner/matching/matching_screen.dart';
 import 'package:projectloner/profile/profile.dart';
 
@@ -28,7 +28,6 @@ class _NavBarState extends State<NavBar> {
       child: ListView(
         children: [
           UserAccountsDrawerHeader(
-
               accountName: Text('Loner') // could be implented later on
               ,
               accountEmail:
